Fix some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/3905



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@16647 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/resbund_cnv.cpp
@@ -44,7 +44,7 @@ ResourceBundle::constructForLocale(const UnicodeString& path,
 {
     char name[300];
 
-    if(path.length() >= sizeof(name)) {
+    if(path.length() >= (int32_t)sizeof(name)) {
         fResource = NULL;
         error = U_ILLEGAL_ARGUMENT_ERROR;
     } else if(!path.isEmpty()) {
